+++ packages/common/src/hooks/index.ts
@@ -26,3 +26,5 @@ export * from './useNotification';
 export * from './useBufferState';
 export * from './useEditModal';
 export * from './useAuthState';
+export * from './useConfirmOnLeaving';
+export * from './useDirtyCheck';
