--- packages/common/src/ui/components/navigation/NavigationPrompt/index.ts
@@ -1,1 +1,0 @@
-export * from './NavigationPrompt';
