+++ packages/common/src/ui/components/navigation/index.ts
@@ -2,4 +2,3 @@ export * from './Breadcrumbs';
 export * from './AppNavLink';
 export * from './ExternalNavLink';
 export * from './Tabs';
-export * from './NavigationPrompt';
