+++ packages/invoices/src/OutboundShipment/DetailView/modals/hooks/useDraftOutboundLines.tsx
@@ -1,9 +1,10 @@
 import React, { useEffect, useState, useCallback } from 'react';
 import {
   generateUUID,
-  isEqual,
   Item,
   StockLine,
+  useConfirmOnLeaving,
+  useDirtyCheck,
   useParams,
 } from '@openmsupply-client/common';
 import { useStockLines } from '@openmsupply-client/system';
@@ -78,7 +79,6 @@ export const createDraftOutboundLine = ({
 interface UseDraftOutboundLinesControl {
   draftOutboundLines: DraftOutboundLine[];
   updateQuantity: (batchId: string, quantity: number) => void;
-  isDirty: boolean;
   isLoading: boolean;
   setDraftOutboundLines: React.Dispatch<
     React.SetStateAction<DraftOutboundLine[]>
@@ -93,13 +93,12 @@ export const useDraftOutboundLines = (
     item?.id ?? ''
   );
   const { data, isLoading } = useStockLines(item?.code ?? '');
-
+  const { isDirty, markDirty } = useDirtyCheck();
   const [draftOutboundLines, setDraftOutboundLines] = useState<
     DraftOutboundLine[]
   >([]);
-  const [initialOutboundLines, setInitialOutboundLines] = useState<
-    DraftOutboundLine[] | undefined
-  >();
+
+  useConfirmOnLeaving(isDirty);
 
   useEffect(() => {
     if (!item) {
@@ -133,27 +132,19 @@ export const useDraftOutboundLines = (
     if (draftOutboundLines?.length === 0) {
       draftOutboundLines.push(createPlaceholderRow(invoiceId, item?.id));
     }
-
-    if (!initialOutboundLines) {
-      setInitialOutboundLines(draftOutboundLines);
-    }
   }, [draftOutboundLines]);
 
   const onChangeRowQuantity = useCallback(
     (batchId: string, value: number) => {
+      markDirty(true);
       setDraftOutboundLines(issueStock(draftOutboundLines, batchId, value));
     },
     [draftOutboundLines]
   );
 
-  const isDirty =
-    !!initialOutboundLines &&
-    !isEqual(initialOutboundLines, draftOutboundLines);
-
   return {
     draftOutboundLines,
     isLoading: isLoading || outboundLinesLoading,
-    isDirty,
     setDraftOutboundLines,
     updateQuantity: onChangeRowQuantity,
   };
